% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_clustering.R
\name{check_clustering}
\alias{check_clustering}
\title{Title}
\usage{
check_clustering(cl_true, cl_estim)
}
\arguments{
\item{cl_true}{Vector of labels}

\item{cl_estim}{Vector of labels}
}
\value{

}
\description{
Title
}
