% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_hclust.R
\name{robust_hclust}
\alias{robust_hclust}
\title{Robust version of hclust}
\usage{
robust_hclust(
  df,
  M = 2,
  Mmax = NULL,
  method = "single",
  M_init = min(floor(nrow(df)/2), 200)
)
}
\arguments{
\item{df}{a data.frame}

\item{M}{number}

\item{Mmax}{Number}

\item{method}{Method to pass to fastcluster::hclust.vector}

\item{M_init}{Number}
}
\value{

}
\description{
Robust version of hclust
}
