% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_blocs.R
\name{simulate_blocs}
\alias{simulate_blocs}
\title{Title}
\usage{
simulate_blocs(n, M = 2, ratio_card_max = 2, epsilon = 0, trunc_eps = TRUE)
}
\arguments{
\item{n}{todo}

\item{M}{todo}

\item{ratio_card_max}{todo}

\item{epsilon}{todo}

\item{trunc_eps}{todo}
}
\value{

}
\description{
Title
}
